+++ src/App.js
@@ -5,6 +5,7 @@ import "./styles.css";
 import Overkill from "./songs/Overkill";
 import Wonderwall from "./songs/Wonderwall";
 import SkyFullOfStars from "./songs/SkyFullOfStars";
+import Help from "./songs/Help";
 import ScrollToTop from "./components/ScrollToTop";
 import {
   toggleScroll,
@@ -35,6 +36,9 @@ const Home = () => (
       </li>
       <li>
         <Link to="/skyfullofstars">Sky Full of Stars - Coldplay</Link>
+      </li>
+      <li>
+        <Link to="/help">Help - Beatles</Link>
       </li>
     </ul>
   </div>
@@ -95,6 +99,14 @@ function App() {
             element={
               <SongLayout>
                 <SkyFullOfStars />
+              </SongLayout>
+            }
+          />
+          <Route
+            path="/help"
+            element={
+              <SongLayout>
+                <Help />
               </SongLayout>
             }
           />