+++ Changelog.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Please stick to the change type wording `Added`, `Changed`, `Removed`, `Fixed`.
 
+## [0.0.3] - 2024-05-27
+
+### Added
+
+- added font resizer
+- added a new song: Help from the Beatles
+- added a new song: Sky full of Stars from Coldplay
+- added chords styling for non-image songs
+
+### Changed
+
+- improved the layout of the song list
+
 ## [0.0.2] - 2024-05-21
 
 ### Added